Move `CEPSearchView` and `CEPSearchViewModel` to its own files
