+++ CEPSearch/CEPiOS/CEP Search/CEPSearchView.swift
@@ -8,43 +8,6 @@
 import SwiftUI
 import CEPSearch
 
-public enum CEPSearchUIComposer {
-    public static func composeView(cepGetter: CEPGetter) -> CEPSearchView {
-        let viewModel = CEPSearchViewModel(cepGetter: cepGetter)
-        
-        // Bind the ViewModel and the View
-        // Keeping ViewModel agnostic
-        let binding = Binding<String>(
-            get: { viewModel.cep },
-            set: { viewModel.cep = $0 }
-        )
-        
-        return CEPSearchView(cep: binding,
-                             viewData: viewModel.viewData,
-                             action: viewModel.fetchCEPDetails)
-    }
-}
-
-final class CEPSearchViewModel {
-    private var cepGetter: CEPGetter
-    var cep: String = .init()
-    
-    // Note: Retrieve localized values
-    var viewData: CEPSearchViewData = {
-        CEPSearchViewData(placeholderText: "Digite o CEP",
-                          buttonText: "Procurar Endereço")
-    }()
-    
-    init(cepGetter: CEPGetter) {
-        self.cepGetter = cepGetter
-    }
-    
-    func fetchCEPDetails(cep: String) async {
-        let _ = try? await cepGetter.getCEPDetails(for: cep)
-        // TODO: Navigation
-    }
-}
-
 struct CEPSearchViewData {
     var placeholderText: String
     var buttonText: String
